requirements.txt
pandas>=1.3.0
numpy>=1.26.0,<2.2.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
requests>=2.26.0
yfinance>=0.2.0
finam-export>=5.1.0
